+++ CHANGES.md
@@ -6,6 +6,7 @@ videojs-record changelog
 
 - Bump required version for videojs-wavesurfer to 2.2.1 for compatibility
   with video.js 6.7.x (#208)
+- Ability to specify bitrate setting in lamejs plugin (#213)
 
 
 2.1.1 - 2018/02/19